feat: one-hot encode context vectors
 - add "num_observations" field to lessons to allow specifying how many (min) observations to take for each problem type. This is an alternative to how many problems to solve, that may better balance the dataset generation so you have approximately the same number of observations for each problem type.

+++ mathzero/model/math_dataset.py
@@ -26,11 +26,11 @@ def make_training_input_fn(examples, batch_size):
 
     output_types = (
         {
-            FEATURE_FWD_VECTORS: tf.int16,
-            FEATURE_BWD_VECTORS: tf.int16,
-            FEATURE_LAST_FWD_VECTORS: tf.int16,
-            FEATURE_LAST_BWD_VECTORS: tf.int16,
-            FEATURE_FOCUS_INDEX: tf.int16,
+            FEATURE_FWD_VECTORS: tf.uint8,
+            FEATURE_BWD_VECTORS: tf.uint8,
+            FEATURE_LAST_FWD_VECTORS: tf.uint8,
+            FEATURE_LAST_BWD_VECTORS: tf.uint8,
+            FEATURE_FOCUS_INDEX: tf.uint8,
             FEATURE_NODE_COUNT: tf.int32,
             FEATURE_MOVE_COUNTER: tf.int32,
             FEATURE_MOVES_REMAINING: tf.int32,
